requirements.txt
langchain-core
langchain-community
langchain_ollama
langchain-openai
langchain_chroma
#python-dotenv
pyppeteer
pypdf2

paddlepaddle        #CPU OCR
paddlepaddle-gpu    #GPU OCR
paddleocr

pymupdf
pdf2image
langdetect
#ollama pull jeffh/intfloat-multilingual-e5-large-instruct:f16

#Speech part
ffmpeg-python
openai-whisper
sounddevice

google-cloud-texttospeech

protobuf==3.20.2 #resolve dependencies issues (paddlepaddle takes priority over google cloud)
